[DEV-10250] Add custom throttle cache for more configurability
This commit generalizes the throttler cache and allows for multiple
local throttling caches in addition to the global one. A "throttling
cache" is a TTL cache of requests to keep track of how many can be
allowed in a given period. This change refactors the existing code to
take an explicit cache as a parameter (relative opaquely) while
maintaining compatiblity with the previous use case of having one global
cache.

Additionally, we bump some deps to get it to run on clojure 1.10.1 and
bump the version number to 0.1.9.

- Bump clojure
- Bump core.cache (and fix a quadratic performance issue
https://clojure.atlassian.net/browse/CCACHE-15)
- Bump midje and lein-midje